+++ src/handlers/startGame.ts
@@ -1,10 +1,7 @@
 import { TypeData } from '../types/enum/typeData.js';
-// import { dataBase } from '../data_base/data_base.js';
 import { AddUserToRoomClient, DataRoom } from '../types/interface/addUser.js';
 import { Socket } from '../types/types/common.js';
-import { userBase } from '../store/userController.js';
-import { roomsBase } from '../store/roomsController.js';
-import { socketBase } from '../store/socketController.js';
+import { roomsBase, userBase, socketBase } from '../store/index.js';
 
 export const addGame = (object: AddUserToRoomClient, socket: Socket) => {
   const { data } = object;
