+++ src/store/index.ts
@@ -1,0 +1,7 @@
+import { gameRoomsBase } from './gameRoomsController.js';
+import { roomsBase } from './roomsController.js';
+import { socketBase } from './socketController.js';
+import { userBase } from './userController.js';
+import { winnersBase } from './winnersController.js';
+
+export { gameRoomsBase, roomsBase, socketBase, userBase, winnersBase };
